Flip after click
Films only become flipable after being clicked

Next film only appears after clicking next film
