+++ really-bad-movie/src/components/Game.jsx
@@ -1,56 +1,5 @@
 import React from "react";
-const MOVIE_URL = "https://api.themoviedb.org/3/";
-const IMAGE_URL = "https://image.tmdb.org/t/p/w500";
-const API_KEY = import.meta.env.VITE_APP_API_KEY;
-
-function fetchFilm(page) {
-  const DISCOVER_URL =
-    MOVIE_URL +
-    `discover/movie?api_key=${API_KEY}&include_video=false&page=${page}`;
-  return fetch(DISCOVER_URL).then((res) => {
-    if (!res.ok) throw new Error("HTTP error");
-    return res.json();
-  });
-}
-
-function FilmTile(props) {
-  const [filmData, setFilmData] = React.useState(null);
-
-  React.useEffect(() => {
-    fetchFilm(props.correct + 1).then((json) => {
-      const film = json.results[props.tileId];
-      setFilmData(film);
-      props.setRating(film.popularity);
-    });
-  }, [props]);
-
-  if (!filmData) return <div>Loading...</div>;
-  return (
-    <>
-      <div className="flip_card">
-        <div className="flip_card_inner">
-          <div className="flip_card_front">
-            <label htmlFor={props.tileId}>
-              Tile {filmData.original_title}
-              <input
-                id={props.tileId}
-                type="image"
-                name={`submit_${props.tileId}`}
-                value={props.tileId}
-                alt={`Poster of ${filmData.original_title}`}
-                src={IMAGE_URL + filmData.poster_path}
-                style={{ height: 500, width: 300 }}
-              />
-            </label>
-          </div>
-          <div className="flip_card_back">
-            <div className="rating">{props.rating.toFixed(0)}</div>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-}
+import FilmTile from "./FilmTile.jsx";
 
 export default function Game(props) {
   const [rating1, setRating1] = React.useState(null);
@@ -65,6 +14,7 @@ export default function Game(props) {
     event.preventDefault();
     console.log(props.lives, props.correct);
     const filmPick = event.nativeEvent.submitter.value;
+
     if (parseInt(filmPick) !== parseInt(higherRating)) {
       props.setLives(props.lives - 1);
     } else {
@@ -72,24 +22,42 @@ export default function Game(props) {
     }
   }
 
+  const nextFilm =
+    5 - props.lives + props.correct !== props.guesses ? (
+      <button
+        onClick={() => {
+          props.setGuesses(props.guesses + 1);
+        }}
+      >
+        Next Film
+      </button>
+    ) : null;
+  if (props.lives === 0) return null;
   //if (!rating1 | !rating2) return <div>Loading...</div>;
   return (
-    <form onSubmit={(event) => handleSubmit(event)}>
-      <div className="flip">
-        <FilmTile
-          tileId={1}
-          correct={props.correct}
-          rating={rating1}
-          setRating={setRating1}
-        />
-        <FilmTile
-          tileId={2}
-          correct={props.correct}
-          rating={rating2}
-          setRating={setRating2}
-        />
-      </div>
-    </form>
+    <>
+      <form onSubmit={(event) => handleSubmit(event)}>
+        <div className="flip">
+          <FilmTile
+            tileId={1}
+            lives={props.lives}
+            correct={props.correct}
+            guesses={props.guesses}
+            rating={rating1}
+            setRating={setRating1}
+          />
+          <FilmTile
+            tileId={2}
+            lives={props.lives}
+            guesses={props.guesses}
+            correct={props.correct}
+            rating={rating2}
+            setRating={setRating2}
+          />
+        </div>
+      </form>
+      {nextFilm}
+    </>
   );
 }
